Address most points from code review
